[dynamo] changed count returns int

+++ vertigo-dynamo-api/src/main/java/io/vertigo/dynamo/kvstore/KVStoreManager.java
@@ -44,7 +44,7 @@ public interface KVStoreManager extends Manager {
 	 * @param collection the collection
 	 * @return count of elements into collection
 	 */
-	long count(String collection);
+	int count(String collection);
 
 	/**
 	 * Adds an element defined by an id in a collection.
